Move data fixtures to top-level conftest.py
The `admin_username` and `admin_password` fixture will find use also in
the functional tests.  This moves them to the top-level `conftest.py` in
order to ensure that they are accessible from all tests.
