+++ tests/conftest.py
@@ -1,0 +1,15 @@
+"""pytest setup and fixtures common for all tests, regardless of suite"""
+
+import os
+
+import pytest
+
+
+@pytest.fixture(scope='session')
+def admin_username():
+    return os.environ.get('ADMINUSERNAME', 'admin')
+
+
+@pytest.fixture(scope='session')
+def admin_password():
+    return os.environ.get('ADMINPASSWORD', 'admin')
